requirements.txt
# ════════════════════════════════════════════════════════
# Kalshi EV Bot — Python Dependencies
# Python 3.11+ required
# ════════════════════════════════════════════════════════

# ── Async HTTP ──────────────────────────────────────────
aiohttp==3.10.11
aiosignal==1.3.1
frozenlist==1.5.0
multidict==6.1.0
yarl==1.18.3

# ── Cryptography (RSA-PSS signing for Kalshi auth) ──────
cryptography==43.0.3
cffi==1.17.1

# ── Fuzzy matching ──────────────────────────────────────
rapidfuzz==3.10.1

# ── Rich terminal UI ────────────────────────────────────
rich==13.9.4
markdown-it-py==3.0.0
mdurl==0.1.2
pygments==2.18.0

# ── Environment variables ───────────────────────────────
python-dotenv==1.0.1

# ── File locking (state persistence) ───────────────────
filelock==3.16.1

# ── HTTP (backtest API calls) ───────────────────────────
requests==2.32.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3

# ── Testing ─────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-mock==3.14.0

# ── Utilities ───────────────────────────────────────────
attrs==24.2.0
typing_extensions==4.12.2
